Move io helpers to io module
